+++ hegi/__init__.py
@@ -16,4 +16,4 @@ __all__ = [
     "SourceMessage",
 ]
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"